+++ pkgs/universal_storage_sync_utils/lib/src/path_validator.dart
@@ -1,9 +1,8 @@
 // ignore_for_file: avoid_catches_without_on_clauses
 
-import 'dart:io';
-
 import 'package:flutter/widgets.dart';
 import 'package:path/path.dart' as p;
+import 'package:universal_io/io.dart';
 
 /// A utility class to validate path properties.
 mixin PathValidator {
@@ -11,20 +10,22 @@ mixin PathValidator {
   ///
   /// It attempts to create and delete a temporary file in the directory.
   /// Returns `true` if successful, `false` otherwise.
-  static Future<bool> isWritable(final String path) async {
+  static bool isWritable(final String path) {
+    if (path.isEmpty) return false;
+
     final tempDir = Directory(path);
     try {
       if (!tempDir.existsSync()) {
         return false;
       }
-      final tempFile = File(
-        p.join(
-          path,
-          '.uss_write_test_${DateTime.now().microsecondsSinceEpoch}',
-        ),
-      );
-      await tempFile.create();
-      await tempFile.delete();
+      File(
+          p.join(
+            path,
+            '.uss_write_test_${DateTime.now().microsecondsSinceEpoch}',
+          ),
+        )
+        ..createSync()
+        ..deleteSync();
       return true;
     } catch (e, stackTrace) {
       debugPrint('Failed to validate path: $e');