+++ pkgs/universal_storage_sync/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0-dev.7] - 2025-10-25
+
+- fixed todo_file_app example to latest changes
+- chore: update interface dependency to 0.1.0-dev.7
+
 ## [0.1.0-dev.6] - 2025-10-18
 
 - chore: from_json_to_json ^0.3.0